Support default-config-dirs with oslo.config
This is a backport from oslo.config master (see
I4df977911539777d1510e8b579375aca5b5f15f4) to support default
config directories.
This is needed because some projects switched to only support WSGI
environment and with that it is no longer possible to add custom
--config-dir parameters when starting the service.

Note: This patch will automatically land in master with the next
oslo.config release. That's why this is not a cherry-pick from master.

Change-Id: If12d8fe7b7a8bdfa47a805db892ef03a94e4a737