Replace patch with sed in XStatic-angular-ui-router template
The setup.py file in the XStatic-angular-ui-router PyPi distribution
differs from the one in the git repo. DLRN uses the latter, so it
fails to build the package.

Changing the patch for a couple of sed
lines achieves the same result, and gives us more flexibility if
needed for future releases (no need to update the patch).

Change-Id: I70ef42d340758b12ed2e2f6821ab6773b3e21555