Add monasca-transform
Include monasca-transform, the aggregation engine for Monasca.
Also requires Apache Spark to be installed separately.

Specifying a Requires: spark == 1.6.3 seems desirable, but there
is currently no package widely available in the community for
installing that version, so the Requires: is left out to allow an
administrator to manually install from the .tgz or find a suitable
package.
To function properly, Monasca Transform requires a number of Spark
libraries (in .jar format) on the local system.  This install will
not change the versions of these .jar files listed in
/etc/monasca-tranform.conf, so an administrator should reference
the .conf file for names and versions to update.
Note that among those .jar files needed is a JDBC driver for connecting
to a MySQL-compatible database instance.  Drizzle-jdbc 1.3 is expected
and configured in monasca-transform.conf.

Different from other deployments, this .spec inserts an
entry_point into the setup.cfg to allow pbr to create a script,
then installs a .service file that points to the resulting
/usr/bin/monasca-transform script.

Note that the name of the monasca_transform directory in the
.tgz includes an _ and thus doesn't match the sname which
has a -. So instead a variable pypi_tar_name is used.

Note there is a hack for the generic_mysql_jar variable.  Rpmlint
falsely identifies this as needing to be replaced with %{_datadir},
when this is really a match string that needs to match the default
value in monasca-transform.conf.  Hack to get around this is to
concatenate two strings.

Future - need to try moving /etc/monasca-transform.conf

Change-Id: I7a1dd98b66a3c74f6ec1ad71394351e328f7bc7b