keystone: disable config generator test
This commit disables the config generator test for the Keystone
package. The test will only work in a tox generated venv but
fail when tests are run by ostestr: It runs oslo-config-generator
in a shell. This in turn requires python-keystone be available
for oslo-config-generator to use at a time when we haven't
even built, let alone installed, a python-keystone package
yet.

Change-Id: I8de2150cd527491db100703dac165ce950f205bf