Remove global-requirements.txt copy
This file has now changed and does not include a minimum
version, since the requirements project split that and moved to
lower-constraints.txt. Thus, it does not make sense for us to keep it
anymore.

On the flip side, we are not using these requirements anymore to
render the required versions of dependencies, so we could be behind
the minimum. We will need to develop tooling for that.

Change-Id: Ib0d74cb3742bdde49104ff07250000fa0606899c