Update osprofiler to 2.3.0
Skip the jaeger tracing tests for now, as neither RDO nor SUSE has a
python jaeger-client package available.

This also requires a patch for python-neutronclient which had been using
a private method that has been removed in this version of osprofiler.
This commit only patches the SUSE version of python-neutronclient to
work around a CI issue, a subsequent patch will ensure the patch applies
to all versions of the package.

Change-Id: I300ded49191723c25c82ad59a53ae7e72128695d