use stestr for os-brick
Also, we are seeing random CI failures due to concurrency. Stestr is
running in parallel by default. See

https://stestr.readthedocs.io/en/latest/MANUAL.html#parallel-testing

This is problematic for time sensitive tests such as this one

https://github.com/openstack/os-brick/blob/master/os_brick/tests/
privileged/test_rootwrap.py#L101

where the thread can be interrupted prior to final time measurement.
This may cause the test to fail randomly. Therefore, we need to
run the tests with the concurrency of 1 to avoid these kinds of
unpredictable failures.

Change-Id: I41e49dab34a92163f7ffc5951cd98a8a88ba02e0