Use basename filter for vmware-nsxlib
There is no tag for what is currently in its pike stable branch in git
and no release on pypi.

renderspec fetch_source and obs service download_files have different
caching behaviour and both are in use. So stable URLs like
vmware-nsx-stable-pike.tar.gz with changing content will not work. The
CI will not update the file that is used to build but the version in the
rendered spec is updated.

Change-Id: I344e45bb147aa5ccbce214a8ef4441e934375cc4