mistral: Do not use basename for source
The sdist tarball needs to be downloaded somehow. Not using
fetch_source() and also not using a full url as Source does not work
in the SUSE CI.

Also remove grpcio from the tooz requirements.txt. It is not needed.
Otherwise, the mistral doc build fails with:

DistributionNotFound: The 'grpcio<1.16.0,>=1.0.0' distribution was \
  not found and is required by tooz

Change-Id: I518552130ecbc0a48cd81b830c6559c92369041b