Fix build without git being installed in build env
Without git being installed, the old (deprecated) pbr integration fails
with

  Warning, treated as error:
  Cannot find git top directory, assuming "."

we did not notice that in the SUSE CI because there was a local hack,
but an actual openSUSE build fails. by switching to sphinx-build
we can avoid this problem, which is more consistent with the rest
of the spec files anyway.

Change-Id: I2dced5316da56e580dddc78d9706ab7562c9257b
(cherry picked from commit 4f65b56dcd1caeca41b8e228d1b0b4ef3cbdce47)