Allow lint runs for a single spec
Currently, a "tox -e lint" run takes a lot of time. But when
developing locally and changing a single spec template, this is
annoying.
So allow a second positional argument for the 3 tools from the tools/
directory which is a string used by the "-name" argument from
"find". With that, it is possible to call lint for a single spec
template with eg:

tox -elint python-cliff

Change-Id: I1ba05ed4e353728d251105fc7ad356d759813481