watcher: add systemd units for watcher services
Currently systemd unit for watcher componenets are missing making it
difficult to manage the lifecycle of these services.
So adding these unit files and be installed during installation.

Also a minor tweak to the permission of the service folder under
sysconf dir to enable reading of the service config files for
these processes.

Change-Id: Id35d7e39429bdf13be95d414481ec6eaf7fce874