Add nova cell initialization to unit file
Without this patch, the nova-conductor service will fail due to not
having an initialized cell0. This initialization needs to happen after
the api_db setup, but before the main db sync, according to the install
docs[1]. This patch adds the necessary step to the nova-api service unit
file and corrects the sequence order.

[1] https://docs.openstack.org/nova/latest/install/controller-install-obs.html#install-and-configure-components

Change-Id: Ic08c2aedd4c9ffb980e61af0dd9d320a89bafa24