monasca-ui: Drop python3 subpackage
Instead of having a subpackage which contains the python module, add
the python module to the main package.
That way we don't have to invent a package name that is not translated
in anyway with pymod2pkg. Currently the subpackage name is
python-horizon-plugin-monasca-ui but that is openSUSE specific. Moving
the module to the main package (that name is translated with
pymod2pkg) solves this.

Change-Id: I812a57cfeecc6992dca7282d55f59739d1aa18d0