neutron: Remove stop action from ovs-cleanup
The neutron-ovs-cleanup script is supposed to be run only on boot.
With ExecStop action being set to execute it as well, it was being
executed also when `systemctl stop` was run.
This was causing connectivity loss when openstack services were
stopped e.g. during upgrade.
Note that many openstack services could be safely stopped without
harm to the underlying services.

Change-Id: I66a536dd8b97f741953bd4ea3999805e2db4a12a