openstacksdk: Temporarily skip unit tests
The openstacksdk unit tests tend to require an obscene amount of
RAM, so a system with 8 GB of RAM might struggle to run them and
use a lot of swap. This is causing many build errors due to timeouts.

Let's skip unit tests for now until this is improved upstream.

Change-Id: I5ef19a206145dda0c7110e4ec8848f89a43e7a52