Fix broken compile options for httpd.bin
This patch changes the build options for httpd.bin. The current
options do not work when the binary is used on cirros. The error
is:

$ ./httpd.bin
-sh: ./httpd.bin: not found

$ ldd httpd.bin
checking sub-depends for '/lib64/libpthread.so.0'
        ld64-uClibc.so.0 => /lib/ld64-uClibc.so.0 (0x7f7b741be000)
        libdl.so.0 => /lib/libdl.so.0 (0x7f7b73fba000)
        libc.so.0 => /lib/libc.so.0 (0x7f7b73d50000)
checking sub-depends for 'not found'
        libpthread.so.0 => /lib64/libpthread.so.0 (0x00000000)
        libc.so.6 => not found (0x00000000)
        /lib/ld64-uClibc.so.0 => /lib/ld64-uClibc.so.0 (0x00000000)

The repo for this tempest plugin provides a different option to
build the httpd.bin

https://github.com/openstack/octavia-tempest-plugin/tree/1.3.0/
octavia_tempest_plugin/contrib/test_server/

Using this option the httpd.bin runs as expected on cirros.

Change-Id: I569d76303030cb1a6dd96a74056a066e7fc876b3
(cherry picked from commit 4cbdd048301ada07601e6d4363bb5b690728ae90)