Adjust all the URLs to point to docs.openstack.org
We had a weird mix of non-existing URLs, and pointing to opendev.org
and to launchpad and other places. Lets put everything point the same
way under https://docs.openstack.org/$project, which redirects to the
newest documentation then automatically.

The other option would be to point to the project navigator, but that
doesn't have all othe components on the granularity that we package
here, and the urls are not stable (require a releases/$releasename
within), so that idea was discarded.

Change-Id: I0ccca433a5e5dbc1082c93d4c5f84ed3fcbed2fe