os-api-ref: add patch to remove sphinx_testing
This patch replaces the sphinx-testing package using the SphinxTestApp
provided by the Sphinx package.

The sphinx-testing package is deprecated and should be replaced with
the sphinx.testing package that's bundled with Sphinx.

this patch is needed to unblock the CI, untill the upstream review is
merged or there is a resolution for this problem.

upstream review: https://review.openstack.org/#/c/860080/

Change-Id: Idf0eead59a3c4f5e281ea7fe0f21d798939f2756