osprofiler: add patch to make compatible with modern elasticsearch
This patch adds default port to Elasticsearch connection url. This
change makes it compatible with modern versions of elasticsearch

this patch is needed to unblock the CI, untill the upstream review is
merged or there is a resolution for this problem.

upstream review: https://review.opendev.org/c/openstack/osprofiler/+/860412

Change-Id: I026b3308f51a53fbff172f88a1fbf488e5ea8206